pyproject.toml
[tool.poetry]
name = "qdamono-server"
version = "2.0.0-alpha.1"
description = ""
authors = [
    "Marcin Białek <bialy.cinek@gmail.com>",
    "Krzysztof Grajda <krzysztof.grajda@protonmail.com>"
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.89.1"}
typer = {extras = ["all"], version = "^0.7.0"}
python-socketio = "^5.7.2"
beanie = "^1.17.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
black = "^23.1.0"
mongomock = "^4.1.2"
mongomock-motor = "^0.0.17"
pytest-asyncio = "^0.20.3"
aiohttp = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "qdamono_server.runner:cli"